v4l4j issue11: Added cast to jlong for C pointers when passing them from JNI code to Java methods expecting "long" argument

+++ v4l4j/branches/issue11/src/v4l4j_DeviceInfo.c
@@ -193,7 +193,7 @@ static void create_formats_object(JNIEnv *e, jobject t, jclass this_class,
 
 
 	//Creates an ImageFormatList
-	obj =  (*e)->NewObject(e, format_list_class, format_list_ctor, (uintptr_t) d);
+	obj =  (*e)->NewObject(e, format_list_class, format_list_ctor, (jlong) (uintptr_t)d);
 	if(obj == NULL){
 		info("[V4L4J] Error creating the format list\n");
 		THROW_EXCEPTION(e, JNI_EXCP, "Error creating the format list");
@@ -288,7 +288,7 @@ JNIEXPORT jobject JNICALL Java_au_edu_jcu_v4l4j_DeviceInfo_doListIntervals(
 		dprint(LOG_V4L4J, "[V4L4L] Creating the frame interval (unsupported)\n");
 		//create the frame interval object
 		frame_intv = (*e)->NewObject(e, frame_intv_class, ctor,
-				3, (uintptr_t) p);
+				3, (jlong) (uintptr_t) p);
 		if(frame_intv == NULL){
 			info("[V4L4J] Error creating FrameInterval object\n");
 			THROW_EXCEPTION(e, JNI_EXCP, \
@@ -300,7 +300,7 @@ JNIEXPORT jobject JNICALL Java_au_edu_jcu_v4l4j_DeviceInfo_doListIntervals(
 		dprint(LOG_V4L4J, "[V4L4L] Creating the frame interval (discrete)\n");
 		//create the frame interval object
 		frame_intv = (*e)->NewObject(e, frame_intv_class, ctor,
-				4, (uintptr_t) p);
+				4, (jlong) (uintptr_t) p);
 		XFREE(p);
 		if(frame_intv == NULL){
 			info("[V4L4J] Error creating FrameInterval object\n");
@@ -313,7 +313,7 @@ JNIEXPORT jobject JNICALL Java_au_edu_jcu_v4l4j_DeviceInfo_doListIntervals(
 		dprint(LOG_V4L4J, "[V4L4L] Creating the frame interval (stepwise)\n");
 		//create the frame interval object
 		frame_intv = (*e)->NewObject(e, frame_intv_class, ctor,
-				5, (uintptr_t) p);
+				5, (jlong) (uintptr_t) p);
 		XFREE(p);
 		if(frame_intv == NULL){
 			info("[V4L4J] Error creating FrameInterval object\n");
